+++ examples/saving/init.py
@@ -23,12 +23,12 @@ sim.saveData(include=['netParams'], filename='out_netParams')
 sim.saveData(include=['net'], filename='out_netInstance')
 
 # save sim config
-sim.saveData(include=['simConfig'], filename='simConfig')
+sim.saveData(include=['simConfig'], filename='out_simConfig')
 
 # save sim output data
-sim.saveData(include=['simData'], filename='simData')
+sim.saveData(include=['simData'], filename='out_simData')
 
 # save network instance with compact conn format (list instead of dict)
-sim.cfg.compactConnFormat = True
+sim.cfg.compactConnFormat = ['preGid', 'sec', 'loc', 'synMech', 'weight', 'delay']
 sim.gatherData()
 sim.saveData(include=['net'], filename='out_netInstance_compact')